django/requirements.txt
certifi==2017.4.17
Django==1.11.2
django-debug-toolbar==1.8
django-split-settings==0.2.5
phonenumbers==8.4.0
psycopg2==2.7.1
pytz==2017.2
sqlparse==0.2.3
virtualenv==15.1.0
djangorestframework==3.6.3
pillow==4.0.0
django-polymorphic
django-cors-headers
django-markdown-app
